requirements.txt
streamlit
faiss-cpu
sentence-transformers
mistralai
unstructured
langchain
python-dotenv